remove transformers that areunecessary, remove mock, move db_handler
